Query reactome for all participating genes in a set of pathways

Get all genes participating in the provided list of reactome pathway identifiers.

query_participating_genes(pathways)

Arguments

pathways
character vector of valid reactome IDs

Value

character vector of ensembl IDs of the genes participating in any of the input pathways.

Examples

query_participating_genes("R-HSA-425407")

#>   [1] "ENSG00000003989" "ENSG00000004809" "ENSG00000004939" "ENSG00000005022"
#>   [5] "ENSG00000007216" "ENSG00000010379" "ENSG00000011083" "ENSG00000017483"
#>   [9] "ENSG00000018280" "ENSG00000021488" "ENSG00000033867" "ENSG00000036565"
#>  [13] "ENSG00000040531" "ENSG00000047457" "ENSG00000050438" "ENSG00000059804"
#>  [17] "ENSG00000064651" "ENSG00000065923" "ENSG00000066230" "ENSG00000070214"
#>  [21] "ENSG00000070915" "ENSG00000072041" "ENSG00000074621" "ENSG00000074803"
#>  [25] "ENSG00000075142" "ENSG00000079215" "ENSG00000080493" "ENSG00000081800"
#>  [29] "ENSG00000084453" "ENSG00000088386" "ENSG00000089060" "ENSG00000089472"
#>  [33] "ENSG00000090020" "ENSG00000091137" "ENSG00000091138" "ENSG00000091664"
#>  [37] "ENSG00000092068" "ENSG00000100156" "ENSG00000100170" "ENSG00000100191"
#>  [41] "ENSG00000100678" "ENSG00000101187" "ENSG00000101200" "ENSG00000101438"
#>  [45] "ENSG00000102100" "ENSG00000102931" "ENSG00000103064" "ENSG00000103257"
#>  [49] "ENSG00000103546" "ENSG00000104635" "ENSG00000104888" "ENSG00000105143"
#>  [53] "ENSG00000105281" "ENSG00000105550" "ENSG00000105641" "ENSG00000106688"
#>  [57] "ENSG00000109667" "ENSG00000110436" "ENSG00000110446" "ENSG00000110628"
#>  [61] "ENSG00000110911" "ENSG00000111181" "ENSG00000111371" "ENSG00000111700"
#>  [65] "ENSG00000112077" "ENSG00000112394" "ENSG00000112473" "ENSG00000112499"
#>  [69] "ENSG00000112759" "ENSG00000113073" "ENSG00000113396" "ENSG00000113504"
#>  [73] "ENSG00000114923" "ENSG00000115194" "ENSG00000115616" "ENSG00000115665"
#>  [77] "ENSG00000115902" "ENSG00000116704" "ENSG00000117394" "ENSG00000117620"
#>  [81] "ENSG00000117834" "ENSG00000118160" "ENSG00000118596" "ENSG00000119899"
#>  [85] "ENSG00000120509" "ENSG00000123643" "ENSG00000124067" "ENSG00000124140"
#>  [89] "ENSG00000124568" "ENSG00000124786" "ENSG00000129353" "ENSG00000130304"
#>  [93] "ENSG00000130876" "ENSG00000130958" "ENSG00000131183" "ENSG00000131389"
#>  [97] "ENSG00000132164" "ENSG00000132677" "ENSG00000132874" "ENSG00000133065"
#> [101] "ENSG00000133460" "ENSG00000134294" "ENSG00000134538" "ENSG00000135740"
#> [105] "ENSG00000136052" "ENSG00000136856" "ENSG00000136868" "ENSG00000137204"
#> [109] "ENSG00000137491" "ENSG00000137860" "ENSG00000137968" "ENSG00000138074"
#> [113] "ENSG00000138079" "ENSG00000138449" "ENSG00000138821" "ENSG00000139155"
#> [117] "ENSG00000139209" "ENSG00000139370" "ENSG00000139514" "ENSG00000139540"
#> [121] "ENSG00000140090" "ENSG00000140199" "ENSG00000140519" "ENSG00000140675"
#> [125] "ENSG00000141424" "ENSG00000141469" "ENSG00000141485" "ENSG00000141526"
#> [129] "ENSG00000141873" "ENSG00000142319" "ENSG00000142494" "ENSG00000143036"
#> [133] "ENSG00000143570" "ENSG00000144136" "ENSG00000144290" "ENSG00000144741"
#> [137] "ENSG00000145217" "ENSG00000145283" "ENSG00000145740" "ENSG00000146411"
#> [141] "ENSG00000146477" "ENSG00000147100" "ENSG00000147606" "ENSG00000147804"
#> [145] "ENSG00000148386" "ENSG00000148942" "ENSG00000149150" "ENSG00000149452"
#> [149] "ENSG00000151012" "ENSG00000151229" "ENSG00000151729" "ENSG00000152683"
#> [153] "ENSG00000154025" "ENSG00000155380" "ENSG00000155465" "ENSG00000155850"
#> [157] "ENSG00000155886" "ENSG00000156222" "ENSG00000157103" "ENSG00000157593"
#> [161] "ENSG00000157765" "ENSG00000158014" "ENSG00000158296" "ENSG00000158467"
#> [165] "ENSG00000158715" "ENSG00000158865" "ENSG00000159216" "ENSG00000160326"
#> [169] "ENSG00000160349" "ENSG00000162383" "ENSG00000162695" "ENSG00000163393"
#> [173] "ENSG00000163406" "ENSG00000163581" "ENSG00000163631" "ENSG00000163817"
#> [177] "ENSG00000164363" "ENSG00000164414" "ENSG00000164638" "ENSG00000164707"
#> [181] "ENSG00000164756" "ENSG00000164889" "ENSG00000165349" "ENSG00000165646"
#> [185] "ENSG00000165794" "ENSG00000165970" "ENSG00000167114" "ENSG00000167703"
#> [189] "ENSG00000168003" "ENSG00000168065" "ENSG00000168575" "ENSG00000169100"
#> [193] "ENSG00000169241" "ENSG00000169359" "ENSG00000170385" "ENSG00000170571"
#> [197] "ENSG00000172270" "ENSG00000173214" "ENSG00000173262" "ENSG00000173930"
#> [201] "ENSG00000174358" "ENSG00000174502" "ENSG00000174640" "ENSG00000174669"
#> [205] "ENSG00000175003" "ENSG00000176463" "ENSG00000177542" "ENSG00000177984"
#> [209] "ENSG00000179520" "ENSG00000180251" "ENSG00000180638" "ENSG00000180773"
#> [213] "ENSG00000181045" "ENSG00000181804" "ENSG00000181830" "ENSG00000181856"
#> [217] "ENSG00000182902" "ENSG00000183023" "ENSG00000183048" "ENSG00000184925"
#> [221] "ENSG00000185052" "ENSG00000186335" "ENSG00000188338" "ENSG00000188467"
#> [225] "ENSG00000188687" "ENSG00000189058" "ENSG00000196517" "ENSG00000196660"
#> [229] "ENSG00000196950" "ENSG00000197119" "ENSG00000197208" "ENSG00000197241"
#> [233] "ENSG00000197375" "ENSG00000197496" "ENSG00000197506" "ENSG00000197818"
#> [237] "ENSG00000197891" "ENSG00000197901" "ENSG00000198246" "ENSG00000198569"
#> [241] "ENSG00000198668" "ENSG00000198689" "ENSG00000198743" "ENSG00000204385"
#> [245] "ENSG00000205060" "ENSG00000206288" "ENSG00000206378" "ENSG00000213465"
#> [249] "ENSG00000215695" "ENSG00000224399" "ENSG00000225697" "ENSG00000226614"
#> [253] "ENSG00000227402" "ENSG00000228263" "ENSG00000229077" "ENSG00000229802"
#> [257] "ENSG00000231479" "ENSG00000232180" "ENSG00000235336" "ENSG00000256870"
#> [261] "ENSG00000262217" "ENSG00000268104" "ENSG00000275744" "ENSG00000276996"
#> [265] "ENSG00000278550" "ENSG00000281165" "ENSG00000281917" "ENSG00000282739"
#> [269] "ENSG00000285096"