Query reactome for involved biological pathways

Get all reactome.org pathways that involve any of the genes provided.

query_reactome_pathways(genes)

Arguments

genes
character vector of valid ensembl gene IDs (no version suffix)

Value

character vector of reactome IDs of the pathways involving at least one of the input genes.

Examples

query_reactome_pathways(c("ENSG00000130203", "ENSG00000189058"))

#>  [1] "R-HSA-1236394" "R-HSA-1251985" "R-HSA-1430728" "R-HSA-162582" 
#>  [5] "R-HSA-174824"  "R-HSA-196854"  "R-HSA-212436"  "R-HSA-2173782"
#>  [9] "R-HSA-2187338" "R-HSA-3000480" "R-HSA-372790"  "R-HSA-381426" 
#> [13] "R-HSA-382551"  "R-HSA-388396"  "R-HSA-392499"  "R-HSA-418594" 
#> [17] "R-HSA-5653656" "R-HSA-597592"  "R-HSA-6806667" "R-HSA-73857"  
#> [21] "R-HSA-74160"   "R-HSA-8864260" "R-HSA-8957275" "R-HSA-8963888"
#> [25] "R-HSA-8963898" "R-HSA-8963899" "R-HSA-8963901" "R-HSA-8964026"
#> [29] "R-HSA-8964043" "R-HSA-8964058" "R-HSA-9006931" "R-HSA-9006934"
#> [33] "R-HSA-9024446" "R-HSA-9029569" "R-HSA-975634"  "R-HSA-425397" 
#> [37] "R-HSA-425407"  "R-HSA-804914"